requirements.txt
django
django-localflavor
validate_docbr
psycopg[binary]
graphene-django
django-webpack-loader
django-bootstrap5
Unidecode
xlsxwriter
pdfminer
tabula-py

#django-auth-ldap
# https://github.com/cgohlke/python-ldap-build/releases
# python_ldap-3.4.4-cp311-cp311-win_amd64.whl

